+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [feat]: `.add()` method for vector addition;
 - [feat]: `.scale()` method for scalar vector multiplication;
 - [feat]: `inverted` property for inverting the vector;
+- [feat]: `.sub()` method for vector subtraction;
 
 ## 0.1.0 (2022-11-23)